+++ src/Views/Shared/_Layout_base.cshtml
@@ -39,7 +39,7 @@
 
     <a href="#main-content" class="govuk-skip-link">@(ViewData["skipLinkMessage"] ?? "Skip to main content")</a>
 
-    <div class="app-cookie-banner">
+    <div class="app-cookie-banner" data-module="cookie-message">
       <p class="app-cookie-banner__message">GOV.UK uses cookies to make the site simpler. <a href="https://www.gov.uk/help/cookies">Find out more about cookies</a</p>
     </div>
